requirements.txt
# Core dependencies
aiogram==3.3.0
aioredis==2.0.1
asyncpg==0.29.0
SQLAlchemy==2.0.23
alembic==1.13.1

# External APIs
openai==1.6.1
anthropic==0.8.1

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Scheduler
APScheduler==3.10.4

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Cache
redis==5.0.1
aiocache==0.12.2

# Utilities
pytz==2023.3
python-dateutil==2.8.2
aiofiles==23.2.1

# HTTP
aiohttp==3.9.1
httpx==0.25.2

# Logging
loguru==0.7.2
colorlog==6.8.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Data processing
pandas==2.1.4
numpy==1.26.2

# Security
cryptography==41.0.7
passlib==1.7.4

# Timezone
tzdata==2023.4
zoneinfo==0.2.1

# Image processing (for future features)
Pillow==10.1.0

# Payment providers (optional)
yookassa==3.0.0

# Optional ML features
scikit-learn==1.3.2
tensorflow-cpu==2.15.0